refactor: CouponCacheManager에서 쿠폰 정보 조회
