+++ src/main/java/coupon/service/MemberCouponService.java
@@ -3,21 +3,16 @@ package coupon.service;
 import static java.util.function.Function.identity;
 import static java.util.stream.Collectors.toMap;
 
-import coupon.cache.CachedCoupon;
 import coupon.domain.Coupon;
 import coupon.domain.Member;
 import coupon.domain.MemberCoupon;
 import coupon.exception.ErrorMessage;
 import coupon.exception.GlobalCustomException;
-import coupon.repository.CachedCouponRepository;
-import coupon.repository.CouponRepository;
 import coupon.repository.MemberCouponRepository;
 import coupon.util.WriterDbReader;
 import java.time.LocalDateTime;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
-import java.util.Optional;
 import java.util.Set;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
@@ -30,9 +25,8 @@ public class MemberCouponService {
     private static final int ISSUE_LIMIT = 5;
 
     private final MemberCouponRepository memberCouponRepository;
-    private final CouponRepository couponRepository;
     private final WriterDbReader writerDbReader;
-    private final CachedCouponRepository cachedCouponRepository;
+    private final CouponCacheManager couponCacheManager;
 
     @Transactional
     public MemberCoupon issue(Member member, Coupon coupon) {
@@ -57,7 +51,7 @@ public class MemberCouponService {
                 .map(MemberCoupon::getCouponId)
                 .collect(Collectors.toSet());
 
-        List<Coupon> coupons = findCouponsWithCache(couponIds);
+        List<Coupon> coupons = couponCacheManager.finAllByIds(couponIds);
         Map<Long, Coupon> couponsById = coupons.stream()
                 .collect(toMap(Coupon::getId, identity()));
 
@@ -66,39 +60,5 @@ public class MemberCouponService {
             memberCoupon.loadCoupon(coupon);
         });
         return issuedCoupons;
-    }
-
-    private List<Coupon> findCouponsWithCache(Set<Long> couponIds) {
-        List<Coupon> cachedCoupons = findCachedCoupons(couponIds);
-
-        Set<Long> cachedCouponIds = cachedCoupons.stream()
-                .map(Coupon::getId)
-                .collect(Collectors.toSet());
-        Set<Long> couponIdsNotInCache = couponIds.stream()
-                .filter(id -> !cachedCouponIds.contains(id))
-                .collect(Collectors.toSet());
-
-        List<Coupon> coupons = new ArrayList<>(cachedCoupons);
-        if (couponIdsNotInCache.size() > 0) {
-            List<Coupon> couponsNotInCache = couponRepository.findAllByIdIn(couponIdsNotInCache);
-            coupons.addAll(couponsNotInCache);
-            saveToCache(couponsNotInCache);
-        }
-        return coupons;
-    }
-
-    private List<Coupon> findCachedCoupons(Set<Long> couponIds) {
-        return couponIds.stream()
-                .map(cachedCouponRepository::findById)
-                .flatMap(Optional::stream)
-                .map(CachedCoupon::getCoupon)
-                .toList();
-    }
-
-    private void saveToCache(List<Coupon> couponsNotInCache) {
-        List<CachedCoupon> cachedCoupons = couponsNotInCache.stream()
-                .map(CachedCoupon::new)
-                .toList();
-        cachedCouponRepository.saveAll(cachedCoupons);
     }
 }
